SPCI: support multiple constituents per memory region.
Previously the memory share mechanism only allowed for a contiguous
set of pages to be shared between two VMs in a single call. More
complex use-cases require non-contiguous pages to be shared between
end-points. This commit creates the necessary infrastructure to allow
non-contiguous pages to be shared in a single call.

Bug: 132420445
Change-Id: Ic9cbd9cb83abe7a55b02e4bfe07ec61e3d1b6822