Enable caching early in assembly.
By doing this early, more code is able to make assumptions about using
the cache so there are fewer areas where manual cache management is
required.

Moving to assembly means less assumptions are made about the state of
the cache before it is enabled. Doing this in C means there are concerns
about the state of the stack if the cache is ever not clean and invalid.

Bug: 141103913
Bug: 139269163
Change-Id: I8ff01c2c2a0c035f147d2d1a8372dbd8bec7ce73