Improve spinlock implementation for ARMv8.0
Current implementation of spinlocks using C11 atomics compiles into an
LDAXR/STXR pair which we observe does not always make forward progress
on Cortex A72 (concurrent_save_restore test livelocks).

This patch inserts a WFE instruction when the LDAXR returns that the
lock is currently taken, pausing the thread until an event is triggered
and easing off the contention. This is observed to help forward progress
on A72 and should also improve power consumption of our spinlocks.

Note that there is still no guarantee of forward progress and threads
can be shown to livelock on the A72 if the loops are extremely tight.
This should be revisited in the future, possibly by leveraging ARMv8.1
LSE atomics.

Test: vcpu_state.concurrent_save_restore
Bug: 141087046
Change-Id: Ic838ba4d287ef4bfecc7dd84c883770d9c445b1d