Move boot params update to the boot_flow module.
Not all boot flows want to be doing updates so leave that decision to
the module responsible for coordinating the boot details.

There is likely a cleaner or more configurable way to acheive this but
the goal for the moment is to isolate the current boot flows that we
have.

Change-Id: Ib0a1d43293e0a70bea897fd4f4695515627bf150