Fix bug when handling EL1 debug register accesses
Certain registers weren't correctly encoded.  Make the encoding scheme similar
to the tables in the Arm Architecture Reference Manual to make it easier to add
new registers as well as spot incorrectly encoded ones.

Fix bug when handling system register accesses with the zero register (XZR) as
the source or the destination.

Log accesses to unhandled registers before they abort, with details on the
encoding of the unhandled register access.

Separate code that is not strictly related to debug registers to facilitate
adding support for performance monitor registers in the future.

Bug: 141452761
Change-Id: Idf717ec56a66ada69d20c590ec2182845afd84b1