Fix return code tracking in hftest.py
Convert DriverRunState from a named tuple to a class because named
tuples are immutable. This resulted in a "thinko" bug where the return
code of a test would always be zero. Note that the return code is used
only to print the epilog of a log and test failure detection was not
affected.

Test: make test fail, check its log
Change-Id: I914f8a95100cfbe207fe1ac8f510c78cbb4fc360