Fix race with checking remaining time on VM timer.
We were assuming that the remaining time on the timer could not be 0
if it had previously been checked not to be pending, but that could
have changed due to time passing since that check was made. To fix
this we get the remaining time once and treat the timer as pending if
the timer is enabled and the remaining time is 0.

Bug: 141469322
Change-Id: I3736a4c1a02acea46e55ce6c41f99e95e42711c0