HCR_EL2 disable traps on Pointer auth usage
Disable traps when the PVM accesses the Pointer Auth key registers or
issues any Pointer Authentication instruction.

The Pointer Auth mechanism holds state in the key registers. Only the
PVM is allowed to use the Pointer Auth functionality for now. This
prevents Hafnium from having to save/restore the key register on a VM
switch.

Change-Id: Ib5299d9e0cd45f8f7bf10569f9d6cb6dea0407b6