Remove deallocation changes in page table commit phase.
The first, preparation, phase in a page table update ensures the page
table has tables allocated to the granularity required for the mapping.
The commit phase can then change the address space without the
possibility of failure and, hence, the potential of a partial update.

Previously, the commit phase would merge empty tables into an absent
entry and free the page. To avoid issues with higher level operations
when the preparation and commit phases details are exposed,
deallocations is removed from the commit phase and an explicit
defragmentation request should be used as needed.

Change-Id: I08b429c6faddf226f66dff1b91f0d217d3a190ca