Convert logically unmapped modes to unmapping.
If an address range is invalid and unowned, it holds no extra
information so can be replaced with an absent entry.

The mapping functions are sometimes used to unmap addresses, for example
during memory sharing. This change means the resulting state of the page
table is the same as though an unmapping function had been used.

This has the side effect of canonicalizing the state for an unmapped
entry.

Change-Id: I1ff35d1dffb9720d5a18c00cd223879fa91c4b24