Inject an Unknown Fault instead of aborting VM on unknown exception
This allows the VM to decide how to handle the fault, rather than being
forcefully aborted.  It also allows us to simplify some of the secondary tests,
whereby the test can trap and observe the exception rather than abort, which
reduces the number of secondary VM services tests need to spawn.

Bug: 140916188
Change-Id: Ib2e98a6507aa15c5d805089cd28851b66c6a6379