Simplify noreturn exception vectors for current EL
All handlers for current EL exceptions panic and never return. Create
noreturn_* variants of exception vector macros without a return path.
Also rename the C handlers to make it clearer in the assembly that these
are not expected to return.

This is done in preparation for inserting speculation barriers after
ERETs, removing the _sp0 handlers which use all the 32 instructions
available in a vector table entry.

Change-Id: If6616686d9a28820ed223e4d3d0e9cdfc3271e52