Do not use `realpath` to get absolute paths
There are two reasons to avoid using `realpath` in our scripts:
(a) It resolves symlinks by default. This can be avoided with a flag.
(b) It is not installed by default on some Linux distros.
Use the `$(cd <dir> && pwd)` trick instead.

This is necessary for checking out Hafnium using repo.

Change-Id: I5dd11d91a55054728723648e5c4192f8f1095b6b