Add optional build flag to check effective image size
Some targets will have a static-size memory carveout dedicated
to the hypervisor. In those cases, we should assert that the Hafnium
image does fit into the allocated memory. Add this as a toolchain arg
'plat_max_image_size' and add a new assertion into check_elf.py. This
way the build will fail if the image is too big.

Change-Id: Ib7311c49b185685aef167988c9aa2416822a6f28