Remove straddling_instruction_unmapped test to fix FVP test failure.
This test was trying to test that an instruction straddling the boundary
between a mapped page and an unmapped page causes a fault due to the
unmapped page. It does on QEMU because QEMU doesn't enforce PC alignment
as the architecture specifies, but on FVP it fails with a PC alignment
fault instead, which is not very interesting to test for.

Change-Id: I90695887fc9913dc842bd535a90d8ef5b30f56c1