Introduce 'stdout' interface
"plat_console" is a build parameter that specifies the console driver
to be used for the given platform. This parameter has been overridden
for all VM toolchains, passing in a HVC-based driver instead.

To keep "plat_console" tied to the platform, introduce "stdout",
which is a platform-agnostic abstraction of an output stream accepting
chars. The default implementation, used by the hypervisor, forwards the
chars to plat_console_putchar(). The implementation in VM toolchains
passes them to the HVC call.

Change-Id: I2316555d9debdb8ce3de4621f8ddbf4b88164d0f