Read device memory regions from the FDT.
If provided, rather than mapping a large chunk of address space as
device memory, allow the specific ranges to be provided and only map
those. Fall back to mapping a large chunk and carving out normal memory
if no device memory regions are provided so QEMU tests continue to work.
These ranges continue to only be mapped to the primary VM for the time
being.

Change-Id: Ia3acf3bb206bf0399ebc3be33d4696cd0450539d