Remove race with test TEAR_DOWN.
If the first CPU finished the loop first and the second CPU received a
message but hasn't yet managed to release the RX buffer, the TEAR_DOWN
handler could find the RX buffer isn't empty and the second CPU meaning
the second CPU would find the RX buffer empty.

To avoid this race, the first CPU needs to wait for the second to
complete before exiting the test case.

Change-Id: Ic0e83385ebdd455e0b3b525d2459bf361645df1f