Add arch_vm_mm_reset()
When running tests on Android dev boards, it is necessary to be able
to break out of the reboot-test loop, boot into the Android Bootloader,
and flash a new system image. The bootloader expects the MMU to be
disabled, so record the initial values of the corresponding system
registers and provide a function to reset them before branching to the
bootloader.

Change-Id: Ic191223e4dc849118e24f2cd7e2ecc113f05b310