Add 'exit' hftest command
A device may need to exit out of the hftest environment, e.g. an Android
device would otherwise always reboot back to hftest and would not be
able to flash a new image.

Add a hftest_device_exit_test_environment() to the hftest_device driver,
where this can be implemented in a board-specific way.

The default PSCI implementation will not support this command and will
print an error message.

Change-Id: I60a92aac21dfbc8b6476f6efe694f05b0967a27d