Support a board-specific manifest overlay for hftests
When an Android boards needs to break out of the hftest test-reboot
loop, it may need more permissions than granted to the primary VM
in the generic hftest manifests. For example, the primary VM may need to
perform SMC calls not included in the SMC whitelist.

To this end, support specifying a manifest overlay in toolchain_args and
use it when generating test initrds.

Change-Id: I5e0de4c25a6ae286b4040825a9d17e471b735134