hftest.py: Add option to skip initial wait for board reboot
Controlling hftests over UART requires some timing synchronization
between host and guest. To this end, hftest.py prompts the user to
restart the board and hit ENTER at the beginning. However, this is
not needed when there is an external script invoking hftest.py on
a series of test suites and only prevents full automation.

Add a command-line argument to skip the initial prompting. It is still
necessary for the user to manually reset the board when a test has
panicked.

Change-Id: I417ad5d2ffd632860cffaa32b920dd9b571411f2