Baremetal VM adapt to SVM
As a transitory phase, the linux test driver must send a direct_req to
initialize the baremetal VM. When Hafnium calls all SPs to initialize
themselves this will no longer be required.

This commit makes the changes required to operate the baremetal VM as a
SVM in secure world. I.e. it creates a suitable init and updated its
UUID (1.1.1.1).

Change-Id: I280d3add3d9364c90d5821edaa15ca7a4bff7b49