FF-A: implement a handler loop for SPMC
When an FF-A ABI is invoked from a partition (virtual FF-A instance)
it enters a handler loop calling ffa_handler such that it can either
handle the FF-A request and resume current or next partition or
forward the call to the normal world. When the SPMC is resumed
from normal world it can process a fresh FF-A request from normal
world. This permits receiving FF-A requests in the SPMC and possibly
return straight to the NWd through SPMD without necessarily resuming
a partition.

Change-Id: Ib00efca90a5419290bd6c9153a6d1c343e8919bb
Signed-off-by: Olivier Deprez <olivier.deprez@arm.com>