Specify initrd image contents as a list of files
The init RAM disk is just a container for various files, and
should be agnostic about their contents (e.g., primary manifest,
kernel, etc...).

Change the build so that a list of files is specified instead of
having the image creator handle different types differently.

This facilitates adding other types of files later, e.g., an FDT
for secondary VMs.

Change-Id: I5ca79dcc5c8810fce0c248201c24f00dd439e392