Unset CHECKPATCH when build script invokes linux driver checkpatch.
The kokoro build script has a step to invoke make checkpatch for the
linux driver. If the CHECKPATCH variable is already set when invoking
the script (non-hermetic builds) then checkpatch for the linux driver
uses the variable already set in the environment which leads to issues
if the checkpatch script does not match the same as the linux directory.
In driver/linux/Makefile the CHECKPATCH variable is only set to a default
path if it's not already set from the caller script environment. This
fix unsets the CHECKPATCH variable from kokoro/build.sh specifically
when invoking linux driver checkpatch.

Change-Id: I428f5955e6c23fb137e1cc16d5cb235f6cda8811
Signed-off-by: Olivier Deprez <olivier.deprez@arm.com>