tests: busy_secondary: enable timer just before ffa_run
The busy_secondary.physical_timer test sometimes fails after
being migrated to jenkins VM. It was not unstable in prior
hosting infrastructure.
The timer might not leave enough time to prepare and send an
FF-A indirect message and for the secondary VM to be scheduled.
The failure happens because the timer fires too early while
still running the PVM.
Also increase the timer timeout value to 3ms from 1ms to be
on the safe side.
Apply the same change to busy_secondary.virtual_timer.

Change-Id: Id139545510c7043adf10c506ac1d394ee0a22c7e
Signed-off-by: Olivier Deprez <olivier.deprez@arm.com>