Handle FF-A calls between worlds consistently with switching VMs.
API functions should set *next to other_world_vcpu to switch to the
other world, rather than NULL. Unrecognised FF-A calls from the other
world (i.e. on the physical FF-A interface) should result in an error
being sent back to the other world, rather than maybe being handled as
another type of call and the error being sent to the last SP to run.

Change-Id: I92e92b38f107f8a15dabd306c77a03b340c04d15
Signed-off-by: Andrew Walbran <qwandor@google.com>