FFA: Adjust 'ffa_memory_permissions_to_mode'
This function is now receiving a default mode. When permissions are not
specified, the data and instruction permissions of the return value
match what set in the default mode.
With this, the receiver's permissions of a memory send operation will
be the same to those of the sender when either the data access or
instruction are not specified.

Change-Id: I6397b4242076ca1f1142288f9424011e429ba15f
Signed-off-by: J-Alves <joao.alves@arm.com>