Tests: Refactoring FVP driver
Refactored FVP driver to decouple tests where hafnium is deployed as
hypervisor, from tests where it is deployed as SPMC.
The class'FVPDriver' was made an abstract class, and kept the base for
'FvpDriverHypervisor' and 'FvpDriverSPMC' derived classes.
Both derived drivers implement the aspect in which they differ, that
are specific to the use-case that they are expected to support.

Change-Id: I6c6532c2871e5cef02b93d72cc725c9eedd11bb8
Signed-off-by: J-Alves <joao.alves@arm.com>
Signed-off-by: Olivier Deprez <olivier.deprez@arm.com>