aarch64: additional permissions granted to Secure Partitions
This changes permits a transition of a legacy TEE to a Secure Partition
on top of the SPMC:
-Disable trap on GIC access: legacy TEEs still directly access the GIC.
This change permits a transitory phase while TEEs are adopting a para-
virtualized interface. This change will be reverted when such interface
is adopted.
-Trap physical interrupts to S-EL1: this is in waiting of handling physical
interrupts at S-EL2 (and managed exit implementation for an NS interrupt
occuring while a SP is running).
-Do not trap WFE/WFI: spinlock usage making use of WFE (low power lock
implementation) triggers a trap to SPMC and attempts to switch to primary.
WFI is possibly used in panic conditions and low power scenarios in a TEE.

Change-Id: Iae5bec2bae685f420ef2d93742c1c402f7aebffc
Signed-off-by: Olivier Deprez <olivier.deprez@arm.com>