Update VM ID print format to hexadecimal
Provided VM IDs are partitioned to non-secure and secure VM IDs (bit 15
clear/set), the VM ID print format looks nicer in the debug logs by
displaying an hexadecimal value rather than a decimal integer (one can
quickly determine if the mentioned ID is a VM or SP and get the ID from
the lower part of the value). This change updates print format strings
to use hex in place of decimal for VM IDs.

Change-Id: I1b957bf79fa824e7759ce88dc15e13cb72764aca
Signed-off-by: Olivier Deprez <olivier.deprez@arm.com>