tests: Fix interruptible_echo test
The test code in the secondary VM contains a loop to handle the case
where ffa_msg_wait() returns with FFA_ERROR_32 with an FFA_INTERRUPTED
error code. The secondary VM never enters the loop, consequently, the
test fails to check that the interrupt actually fired. Moreover, the
corresponding primary VM code in deliver_interrupt_and_message does not
have code to handle the case where the secondary VM yields in the loop,
and to resume the secondary VM again.
This patch removes the loop and moves the check for interrupt assertion
to run every time.

Signed-off-by: Raghu Krishnamurthy <raghu.ncstate@icloud.com>
Change-Id: I3a3298317820eb7292452a6fb123d09a7ea7a4d3