FIX: Consistency handling SMC32/64 IDs
The function 'ffa_handler' was discarding the bit of the used
calling convention out of the func id, whereas in handling FF-A features
the calling convention bit was being considered.
Inconsistency relevant for the case of the interface FFA_RXTX_MAP [1],
whose support for 32-bit calling convention is not yet implemented.

[1] https://lists.trustedfirmware.org/pipermail/hafnium/2020-July/000063.html

Change-Id: I8ec027a6e65dcf6d1af297a6808554ecff967c3b
Signed-off-by: J-Alves <joao.alves@arm.com>