Tests: Remove alignment checking for service VM's
When running memory sharing test suite on AEM FVP, there were test
failures. Debug showed that all the test failures were caused due to
pointers to ffa_composite_memory_region structures during memory sharing
not being size aligned. Due to this, ffa_memory_region_constituent
structure in the ffa_composite_memory_region structure's were also
unaligned. sctlr_el1.A bit is set for all test VM's due to which there
were faults being generated when the above structures were accessed.
To fix the issue there are couple of options: convert all accesses to
the above structures to use memcpy or set sctlr_el1.A bit to 0 for the
test VM's. Note that making the structure itself aligned is not enough
since we can allocate memory that is not aligned and assign it to a
pointer to the structure.
Setting sctlr_el1.A bit appears the easier and clearner option since the
change is small and there is not much benefit in enforcing aligned
access in test code.
It is still unclear why qemu does not have an issue but FVP does.
However, this change was tested on both FVP and qemu to make sure there
were no failures.

Change-Id: I2fa8a974856a8f73608d8257a8ef29f070b23600
Signed-off-by: Raghu Krishnamurthy <raghu.ncstate@icloud.com>