VHE: Move hcr_el2 from lazy saved region to volatile region.
This patch moves hcr_el2 from the list of registers that are saved
lazily to the volatile region, so that it is saved and restored on every
entry and exit into EL2. This is to prepare for future changes that
require enabling hcr_el2.TGE at every entry into EL2 so that hafnium can
execute in "host" mode when FEAT_VHE is enabled and available.

Change-Id: I1e4570ee07df70eb1b608ad8b26a791bf9a50f2a
Signed-off-by: Raghu Krishnamurthy <raghu.ncstate@gmail.com>