VHE: Enable HCR_EL2.TGE on all entries into EL2
This patch enables HCR_EL2.TGE when VHE is enabled and available on the
platform. With this, hafnium runs in "host" mode with HCR_EL2.{E2H,TGE}
= {1, 1}, similar to the linux kernel when VHE is enabled. To be
precise, TGE enables trapping EL1 exceptions to EL2.
This patch prepares hafnium for future changes that can enable execution
of EL0 partitions (primarily S-EL0 partitions), which is allowed by the
FF-A 1.0 spec.

Change-Id: If684f04d1c83458d6cbdf1a3dd17225ba1ad03c2
Signed-off-by: Raghu Krishnamurthy <raghu.ncstate@gmail.com>